Update torch.call_function constant fold through logic
